refactor: move getMaxAgeFromExpiresAt function to config and update imports
